Config command line argument passed into __main__.
