+++ py_pRF_mapping/analysis/utilities.py
@@ -170,3 +170,20 @@ def crt_hrf(varNumVol, varTr):
     vecHrf = np.divide(vecHrf, np.max(vecHrf))
 
     return vecHrf
+
+
+class cls_set_config(object):
+    """
+    Set config parameters from dictionary into local namespace.
+
+    Parameters
+    ----------
+    dicCnfg : dict
+        Dictionary containing parameter names (as keys) and parameter values
+        (as values). For example, `dicCnfg['varTr']` contains a float, such as
+        `2.94`.
+    """
+
+    def __init__(self, dicCnfg):
+        """Set config parameters from dictionary into local namespace."""
+        self.__dict__.update(dicCnfg)
